Spring v3

+++ speedy-jpa-impl/src/test/java/com/github/silent/samurai/speedy/jpa/impl/data/Product.java
@@ -13,6 +13,6 @@ public class Product {
     String id;
     String name;
     String category;
-    Integer cost;
+    Long cost;
 
 }
